Reverting all changes to SbHash. They will not be part of 3.1.

+++ src/misc/SoDBP.h
@@ -57,7 +57,7 @@ public:
   void * userdata;
 };
 
-typedef SbHash<uint32_t, int16_t> UInt32ToInt16Map;
+typedef SbHash<int16_t, uint32_t> UInt32ToInt16Map;
 
 // *************************************************************************
